requirements.txt
### General requirements ###
spotipy
python-decouple
PyQt6